First flights (#1)
* Add data from first flights

Add cleaned first flight data from the CCE dataset. It includes first flights of 9 pigeons from the Solo condition and 10 pigeons from the first generation of the experimental condition.

* Update code to process and fit first flights

Change code to process the first flights, and fit variables for the goal and continuity components (excluding memory and social components).
Also a minor change to preamble.R to specify the mirror for installing new packages.